+++ src/main/java/bowling/domain/frame/FrameStatus.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 import java.util.List;
