+++ src/test/java/bowling/domain/frame/DownedPinTest.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 import bowling.bowlingexception.InvalidDownedPinNumberException;
 import org.junit.jupiter.api.DisplayName;
