+++ src/test/java/bowling/domain/game/BowlingGameTest.java
@@ -1,5 +1,6 @@
-package bowling.domain;
+package bowling.domain.game;
 
+import bowling.domain.frame.DownedPin;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
